requirements.txt
# Hosting requirements
fastapi
starlette
uvicorn[standard]

# Dependency of FastAPI
pydantic
websockets

# Database handling
nanoid
pymongo[zstd,encryption]

# Password hashing
argon2-cffi

# TOML parsing
# Module is built in on Python 3.11+
tomli; python_version < "3.11"
